Moved previous /character to /character/all
